requirements.txt
Faker==37.11.0
mysql-connector-python==9.4.0
tzdata==2025.2
